LayoutTests/imported/w3c/web-platform-tests/fetch/api/redirect/redirect-mode-expected.txt


FAIL Redirect 301 in "error" mode  assert_unreached: Should have rejected. Reached unreachable code
PASS Redirect 301 in "follow" mode  
FAIL Redirect 301 in "manual" mode  assert_equals: Response's status is 0 expected 0 but got 200
FAIL Redirect 302 in "error" mode  assert_unreached: Should have rejected. Reached unreachable code
PASS Redirect 302 in "follow" mode  
FAIL Redirect 302 in "manual" mode  assert_equals: Response's status is 0 expected 0 but got 200
FAIL Redirect 303 in "error" mode  assert_unreached: Should have rejected. Reached unreachable code
PASS Redirect 303 in "follow" mode  
FAIL Redirect 303 in "manual" mode  assert_equals: Response's status is 0 expected 0 but got 200
FAIL Redirect 307 in "error" mode  assert_unreached: Should have rejected. Reached unreachable code
PASS Redirect 307 in "follow" mode  
FAIL Redirect 307 in "manual" mode  assert_equals: Response's status is 0 expected 0 but got 200
FAIL Redirect 308 in "error" mode  assert_unreached: Should have rejected. Reached unreachable code
PASS Redirect 308 in "follow" mode  
FAIL Redirect 308 in "manual" mode  assert_equals: Response's status is 0 expected 0 but got 200
